backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Presidio dependencies
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
presidio-image-redactor==0.0.57

# Additional anonymization libraries (from peterhubina/anonymization)
spacy>=3.4.4,<4.0.0
pytesseract>=0.3.10
PyMuPDF==1.23.9
python-docx>=0.8.11
opencv-python>=4.8.0
azure-ai-formrecognizer>=3.3.0,<4.0.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Async support
aiofiles>=23.2.0
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
